pyproject.toml
[tool.poetry]
name = "pythonproject11"
version = "0.1.0"
description = ""
authors = ["Student <andreevdamir266@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-cors-headers = "^4.7.0"
setuptools = "^76.1.0"
djangorestframework-simplejwt = "^5.5.0"
celery = "^5.4.0"
redis = "^5.2.1"
python-telegram-bot = "^22.0"
drf-yasg = "^1.21.10"
serializers = "^0.2.4"
psycopg2-binary = "^2.9.10"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
